requirements.txt
django
djangorestframework
mysqlclient
pyjwt